Replace get_pollution_values
This function has been replaced with 'pollution_series' which returns a
Pandas Series of pollution values for each node. This has the advantage
that it can be queried with the names of nodes (like a dictionary). It
is also easy to cast to a list using the list function.

This has removed some duplicate code